requirements.txt
rpaframework
selenium
openpyxl
